backend/requirements.txt
aerich==0.7.2
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.2
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==4.2.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
asyncssh==2.14.2
attrs==23.2.0
Babel==2.11.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpayments==1.6.3
cryptography==42.0.1
dictdiffer==0.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.3.1
h11==0.14.0
idna==3.6
iso8601==1.1.0
loguru==0.7.2
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.4
nest-asyncio==1.6.0
phonenumbers==8.13.29
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.4
requests==2.31.0
rsa==4.9
secure==0.3.0
semver==3.0.2
sentry-sdk==1.39.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.27.0.post1
validators==0.22.0
wait-for-it==2.2.2
watchfiles==0.21.0
yarl==1.8.1
